challenges: Cache the number of challenges
getattr needs the number of subdirectories in order to build the struct
stat. The default behavior is to refresh the files to get them, which
would download all the challenge lists from all the categories just to
do an ls -l in the 'challenges' directory.

This patch caches get the number of challenges per category from the
category list. This cached value is also updated every time the
challenge list is downloaded for a category.

Signed-off-by: Celelibi <celelibi@gmail.com>